Refactor tasks into separate files with class-based self-registration
Move task implementations from task_registry.cpp into engine/src/tasks/
with one file per task. Each task is now a class with static spec()
and run() methods, and self-registers via TaskRegistrar<T> template.

Structure:
- engine/src/tasks/viewer_follow.cpp
- engine/src/tasks/viewer_fetch_cached_recommendation.cpp
- engine/src/tasks/take.cpp
- engine/src/tasks/vm.cpp
- engine/src/tasks/filter.cpp
- engine/src/tasks/concat.cpp

Benefits:
- Cleaner separation of concerns (registry infra vs task logic)
- Each task is self-contained with spec and implementation together
- Easier to add new tasks (just create file with class + registrar)
- task_registry.cpp now contains only registry infrastructure
