+++ engine/CMakeLists.txt
@@ -31,12 +31,23 @@ set(CATCH_BUILD_TESTING OFF CACHE BOOL "" FORCE)
 
 FetchContent_MakeAvailable(json cli11 Catch2)
 
+# Task source files (self-registering via static TaskRegistrar)
+set(TASK_SOURCES
+  src/tasks/viewer_follow.cpp
+  src/tasks/viewer_fetch_cached_recommendation.cpp
+  src/tasks/take.cpp
+  src/tasks/vm.cpp
+  src/tasks/filter.cpp
+  src/tasks/concat.cpp
+)
+
 add_executable(rankd
   src/main.cpp
   src/plan.cpp
   src/executor.cpp
   src/task_registry.cpp
   src/output_contract.cpp
+  ${TASK_SOURCES}
 )
 
 target_include_directories(rankd PRIVATE ${CMAKE_SOURCE_DIR}/include)
@@ -54,6 +65,7 @@ add_executable(rankd_tests
   tests/test_pred_eval.cpp
   src/task_registry.cpp
   src/output_contract.cpp
+  ${TASK_SOURCES}
 )
 
 target_include_directories(rankd_tests PRIVATE ${CMAKE_SOURCE_DIR}/include)
@@ -70,6 +82,7 @@ add_executable(concat_tests
   src/executor.cpp
   src/task_registry.cpp
   src/output_contract.cpp
+  ${TASK_SOURCES}
 )
 
 target_include_directories(concat_tests PRIVATE ${CMAKE_SOURCE_DIR}/include)
